Move misplaced classes to correct package
While at it, rename Spec to IntegrationSpec.
