+++ src/java/com/unifina/security/MySecurityManager.java
@@ -1,6 +1,5 @@
 package com.unifina.security;
 
-
 public class MySecurityManager extends SecurityManager {
 	@Override
 	public void checkAccess(ThreadGroup g) {
